--- application/web-java-openliberty/src/main/java/webapp/controller/WlpWebJavaSpringApplication.java
@@ -1,15 +1,0 @@
-package webapp.controller;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-import org.springframework.stereotype.Controller;
-
-@Controller
-@SpringBootApplication
-public class WlpWebJavaSpringApplication extends SpringBootServletInitializer {
-
-    public static void main(String[] args) {
-        SpringApplication.run(WlpWebJavaSpringApplication.class, args);
-    }
-}
