+++ application/web-java-openliberty/src/main/java/webapp/tier/resource/JaxRsApplication.java
@@ -1,0 +1,9 @@
+package webapp.tier.resource;
+
+import javax.ws.rs.ApplicationPath;
+import javax.ws.rs.core.Application;
+
+@ApplicationPath("api")
+public class JaxRsApplication extends Application {
+
+}
