+++ application/web-java-openliberty/src/main/java/webapp/tier/resource/MemcachedResource.java
@@ -1,28 +1,26 @@
-package webapp.controller;
+package webapp.tier.resource;
 
 import javax.ws.rs.Consumes;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
 import webapp.tier.cache.MemcachedService;
 
-@RestController
-@RequestMapping("/memcached")
+@Path("/memcached")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-public class MemcachedController {
+public class MemcachedResource {
 
 	MemcachedService createMemcachedService() {
 		return new MemcachedService();
 	}
 
-	@PostMapping("/set")
+	@POST
+	@Path("/set")
 	public Response set() {
 		MemcachedService svc = createMemcachedService();
 		try {
@@ -33,7 +31,8 @@ public class MemcachedController {
 		}
 	}
 
-	@GetMapping("/get")
+	@GET
+	@Path("/get")
 	public Response get() {
 		MemcachedService svc = createMemcachedService();
 		try {
