+++ application/web-java-openliberty/src/main/java/webapp/tier/resource/RabbitmqResource.java
@@ -1,19 +1,16 @@
-package webapp.controller;
+package webapp.tier.resource;
 
 import javax.ws.rs.Consumes;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
 import webapp.tier.mq.RabbitmqService;
 
-@RestController
-@RequestMapping("/rabbitmq")
+@Path("/rabbitmq")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 public class RabbitmqResource {
@@ -22,7 +19,8 @@ public class RabbitmqResource {
 		return new RabbitmqService();
 	}
 
-	@PostMapping("/put")
+	@POST
+	@Path("/put")
 	public Response put() {
 		RabbitmqService svc = createRabbitmqService();
 		try {
@@ -33,7 +31,8 @@ public class RabbitmqResource {
 		}
 	}
 
-	@GetMapping("/get")
+	@GET
+	@Path("/get")
 	public Response get() {
 		RabbitmqService svc = createRabbitmqService();
 		try {
@@ -44,7 +43,8 @@ public class RabbitmqResource {
 		}
 	}
 
-	@PostMapping("/publish")
+	@POST
+	@Path("/publish")
 	public Response publish() {
 		RabbitmqService svc = createRabbitmqService();
 		try {
