+++ application/web-java-openliberty/src/main/java/webapp/tier/resource/RedisResource.java
@@ -1,19 +1,16 @@
-package webapp.controller;
+package webapp.tier.resource;
 
 import javax.ws.rs.Consumes;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
 import webapp.tier.cache.RedisService;
 
-@RestController
-@RequestMapping("/redis")
+@Path("/redis")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 public class RedisResource {
@@ -22,7 +19,8 @@ public class RedisResource {
 		return new RedisService();
 	}
 
-	@PostMapping("/put")
+	@POST
+	@Path("/set")
 	public Response set() {
 		RedisService svc = createRedisService();
 		try {
@@ -33,7 +31,8 @@ public class RedisResource {
 		}
 	}
 
-	@GetMapping("/get")
+	@GET
+	@Path("/get")
 	public Response get() {
 		RedisService svc = createRedisService();
 		try {
@@ -44,7 +43,8 @@ public class RedisResource {
 		}
 	}
 
-	@PostMapping("/publish")
+	@POST
+	@Path("/publish")
 	public Response publish() {
 		RedisService svc = createRedisService();
 		try {
