+++ application/web-java-openliberty/src/main/liberty/config/server.xml
@@ -1,0 +1,23 @@
+<server description="Sample Liberty server">
+	<featureManager>
+		<feature>servlet-4.0</feature>
+		<feature>localConnector-1.0</feature>
+		<feature>microProfile-3.3</feature>
+		<feature>monitor-1.0</feature>
+		<feature>jndi-1.0</feature>
+		<feature>jms-2.0</feature>
+	</featureManager>
+
+	<variable name="default.http.port" defaultValue="9080" />
+	<variable name="default.https.port" defaultValue="9443" />
+
+	<webApplication location="openliberty.war"
+		contextRoot="/" />
+	<mpMetrics authentication="false" />
+
+	<httpEndpoint host="*" httpPort="${default.http.port}"
+		httpsPort="${default.https.port}" id="defaultHttpEndpoint" />
+
+	<variable name="io_openliberty_guides_system_inMaintenance"
+		value="false" />
+</server>
