+++ application/web-java-openliberty/src/main/webapp/index.html
@@ -1,0 +1,39 @@
+<!--
+  Copyright (c) 2016, 2020 IBM Corp.
+
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+-->
+<html>
+<body>
+    <h1>Welcome to your Open Liberty application</h1>
+    <p>Open Liberty is a lightweight open framework for building fast and efficient cloud-native Java microservices. Find out more at <a href="http://openliberty.io/" target="_blank" rel="noopener noreferrer">openliberty.io</a>.</p>
+<div>
+    <h2>Eclipse MicroProfile</h2>
+    <p>
+        The <a href="https://microprofile.io/" target="_blank" rel="noopener noreferrer">Eclipse MicroProfile project</a> is an open community with the aim of optimizing enterprise Java for a microservices architecture. 
+        MicroProfile evolves with guidance from the community.
+    </p>
+    <p>
+        If you want to share your thoughts, you can post straight to the
+        <a href="https://groups.google.com/forum/#!forum/microprofile" target="_blank" rel="noopener noreferrer">MicroProfile Google group</a>.
+    </p>
+    <p>
+        For more information about the features used in this application, see the Open Liberty documentation:
+        <ul>
+            <li><a href="https://openliberty.io/docs/ref/feature/#microProfile-3.3.html" target="_blank" rel="noopener noreferrer">MicroProfile 3.3</a></li>
+            <li><a href="https://openliberty.io/docs/ref/feature/#jaxrs-2.1.html" target="_blank" rel="noopener noreferrer">Java RESTful Services 2.1</a></li>
+        </ul>
+    </p>
+</div>
+</body>
+</html>
